Fixed bug on o3DS, std::bad_alloc when block is u16. - this means that for now, we can only operarte on 0 - 255  values.
In the past, this worked, likely indicating, that the o3DS is running out of heap space...

+++ include/recraft/blocks/CT_Block.h
@@ -8,7 +8,7 @@
 #include "mclib/block/Block.h"
 #include "mclib/inventory/Slot.h"
 
-typedef u16 Block;
+typedef u8 Block; // TODO: Convert to u16 once the o3DS hardware supports it
 
 //TODO: Instead of ENUM IDs, use the "flattened" strings
 enum { Block_Air,
